requirements.txt
# 数据处理和科学计算
numpy>=1.21.0
pandas>=1.3.0

# 数据库
duckdb>=0.9.0

# 文本处理和搜索
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
python-magic-bin>=0.4.14
xmltodict>=0.13.0
openai>=1.0.0

# 进度条显示
tqdm>=4.62.0

# 测试
pytest>=6.2.0

# 开发工具
black>=21.5b2
flake8>=3.9.0

# 新增依赖
python-dotenv 